LPS-131631 Refactoring
